requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.2
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
